+++ packages/cardano-services/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.26.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/cardano-services@0.26.4...@cardano-sdk/cardano-services@0.26.5) (2024-02-29)
+
+**Note:** Version bump only for package @cardano-sdk/cardano-services
+
 ## [0.26.4](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/cardano-services@0.26.3...@cardano-sdk/cardano-services@0.26.4) (2024-02-28)
 
 **Note:** Version bump only for package @cardano-sdk/cardano-services