+++ packages/e2e/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.32.8](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/e2e@0.32.7...@cardano-sdk/e2e@0.32.8) (2024-02-29)
+
+**Note:** Version bump only for package @cardano-sdk/e2e
+
 ## [0.32.7](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/e2e@0.32.6...@cardano-sdk/e2e@0.32.7) (2024-02-28)
 
 **Note:** Version bump only for package @cardano-sdk/e2e