+++ packages/governance/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.15](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/governance@0.9.14...@cardano-sdk/governance@0.9.15) (2024-02-29)
+
+**Note:** Version bump only for package @cardano-sdk/governance
+
 ## [0.9.14](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/governance@0.9.13...@cardano-sdk/governance@0.9.14) (2024-02-28)
 
 **Note:** Version bump only for package @cardano-sdk/governance