+++ packages/projection/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.8](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection@0.11.7...@cardano-sdk/projection@0.11.8) (2024-02-29)
+
+**Note:** Version bump only for package @cardano-sdk/projection
+
 ## [0.11.7](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection@0.11.6...@cardano-sdk/projection@0.11.7) (2024-02-28)
 
 **Note:** Version bump only for package @cardano-sdk/projection