+++ packages/util-rxjs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.7](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/util-rxjs@0.7.6...@cardano-sdk/util-rxjs@0.7.7) (2024-02-29)
+
+**Note:** Version bump only for package @cardano-sdk/util-rxjs
+
 ## [0.7.6](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/util-rxjs@0.7.5...@cardano-sdk/util-rxjs@0.7.6) (2024-02-28)
 
 **Note:** Version bump only for package @cardano-sdk/util-rxjs