+++ packages/wallet/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.34.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/wallet@0.34.4...@cardano-sdk/wallet@0.34.5) (2024-02-29)
+
+**Note:** Version bump only for package @cardano-sdk/wallet
+
 ## [0.34.4](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/wallet@0.34.3...@cardano-sdk/wallet@0.34.4) (2024-02-28)
 
 ### Features